Prototype of First Stage Complete
First prototype of the first stage is complete, proving the concept is possible, I think.

Things that work: Fixed jittering issue with control action (was missing an abs()), added pulldown resistors to circuit but not sure if they do anything.

TODO:
Figure out how to make a good controller. Current implementation is a bunch of arbitrary values and bad nonlinearities/conditions for when the integral action kicks in.
Improve fit and finish by adding inserts to better clamp to the M16 rod.
Do a whole bunch of calibration (and figure out exactly what calibration you need to do).
Tidy up the wiring
Finish up the CAD for second stage and start printing it